% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_experiments.R
\name{compute_ci}
\alias{compute_ci}
\title{Compute the confidence intervals (CI) for the accuracy using the quantile method.}
\usage{
compute_ci(avec, alpha = 5)
}
\arguments{
\item{avec}{A vector with accuracy values.}
}
\value{
The results with the CIs added.
}
\description{
Compute the confidence intervals (CI) for the accuracy using the quantile method.
}

