% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cpermute}
\alias{cpermute}
\title{Produces a permutation of integers 1:length(vec) such that the
permutation leaves vec unchanged, i.e., the permutations are within
class labes defined by vec.}
\usage{
cpermute(vec, replace = FALSE)
}
\arguments{
\item{vec}{Vector defining the permutations}

\item{replace}{Should the sampling be with or without
replacement. Default is FALSE.}
}
\value{
A permutation.
}
\description{
Produces a permutation of integers 1:length(vec) such that the
permutation leaves vec unchanged, i.e., the permutations are within
class labes defined by vec.
}

