% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{defactor}
\alias{defactor}
\title{Defactor a numeric.}
\usage{
defactor(x)
}
\arguments{
\item{x}{A numeric factor to convert to a true numeric.}
}
\value{
A numeric.
}
\description{
Defactor a numeric factor to a true numeric.
}

