% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_experiments.R
\name{do_experiment_synthetic}
\alias{do_experiment_synthetic}
\title{Perform experiments on synthetic datasets.
                                      #
Perform experiments on synthetic datasets.}
\usage{
do_experiment_synthetic(classifier_list, datapath, savepath, R = NULL,
  Rmin = NULL, Rmax = NULL, z = NULL, alpha = NULL,
  prune_singletons = FALSE)
}
\arguments{
\item{classifier_list}{List of classifiers}

\item{datapath}{Directory with the data files.}

\item{savepath}{Path where to save the resutls.}

\item{R}{Number of samples to use. Default is\code{NULL} in which
case R = 1 + (1 / alpha) will be used by
\code{analyze_dataset}.}

\item{Rmin}{Ninimum number of replications to use for calculating p-values.}

\item{Rmax}{Maximum number of replications to use for calculating p-values.}

\item{z}{Parameter for confidence band for p-values. Use 1.96 for 95 percent, 2.25 for 97.5 percent and 2.6 for 99 percent. Default is 2.6.}

\item{alpha}{Significance level (default is 0.05).}

\item{prune_singletons}{Should singletons be pruned. Default is \code{TRUE}.}
}
\value{
Nothing.
}
\description{
Perform experiments on synthetic datasets.
                                      #
Perform experiments on synthetic datasets.
}

