% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_experiments.R
\name{do_experiment_uci}
\alias{do_experiment_uci}
\title{Perform experiments on UCI datasets (in arff format).
                                      #
Perform experiments on UCI datasets (in arff format).}
\usage{
do_experiment_uci(dataset_list, classifier_list, datapath, savepath, R = NULL,
  Rmin = NULL, Rmax = NULL, z = NULL, alpha = NULL,
  prune_singletons = FALSE, seed = 42, seed_traintest = 42,
  early_stopping = FALSE)
}
\arguments{
\item{dataset_list}{List of datasets}

\item{classifier_list}{List of classifiers}

\item{datapath}{Directory with the data files.}

\item{savepath}{Path where to save the resutls.}

\item{R}{Number of samples to use. Default is\code{NULL} in which
case R = 1 + (1 / alpha) will be used by
\code{analyze_dataset}.}

\item{Rmin}{Ninimum number of replications to use for calculating p-values.}

\item{Rmax}{Maximum number of replications to use for calculating p-values.}

\item{z}{Parameter for confidence band for p-values. Use 1.96 for 95 percent, 2.25 for 97.5 percent and 2.57 for 99 percent. Default is 2.57.}

\item{alpha}{Significance level (default is 0.05).}

\item{prune_singletons}{Should singletons be pruned. Default is \code{TRUE}.}

\item{seed}{Random seed, default is 42.}

\item{seed_traintest}{Random seed used for splitting the data into training and testing sets, default is 42.}

\item{early_stopping}{Use early stopping. Default is \code{FALSE} in which case Rmin samples are used to calculate p-values. If \code{TRUE} at least Rmin and at most Rmax values are used.}
}
\value{
Nothing.
}
\description{
Perform experiments on UCI datasets (in arff format).
                                      #
Perform experiments on UCI datasets (in arff format).
}

