% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_experiments.R
\name{get_dataset_properties}
\alias{get_dataset_properties}
\title{Get properties of a dataset.
                                      #
Get properties of a dataset.}
\usage{
get_dataset_properties(dataset)
}
\arguments{
\item{dataset}{The dataset (a list).}
}
\value{
The dataset structure with a new list element \code{properties} added containing properties of the dataset.
}
\description{
Get properties of a dataset.
                                      #
Get properties of a dataset.
}

