% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{get_full_treelist}
\alias{get_full_treelist}
\title{Return full list of trees.}
\usage{
get_full_treelist(res, alpha = 0.05)
}
\arguments{
\item{res}{A results structure, i.e., the output from
\code{analyze_dataset}.}

\item{alpha}{The significance level for filtering (showing) trees
only above this threshold. The default is 0.05.}
}
\value{
Full list of trees.
}
\description{
Return full list of trees, including trees for which p < alpha.
}

