% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{get_k_tree}
\alias{get_k_tree}
\title{Get the k:th tree from a list of trees.}
\usage{
get_k_tree(treelist, k)
}
\arguments{
\item{treelist}{List of trees.}

\item{k}{Index of tree to get.}
}
\value{
The k:th tree.
}
\description{
Get the k:th tree from a list of trees.
}

