% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_experiments.R
\name{list_datasets}
\alias{list_datasets}
\title{List datasets (arff-files) in a given path.
                                      #
List datasets (arff-files) in a given path.}
\usage{
list_datasets(datapath)
}
\arguments{
\item{datapath}{A path.}
}
\value{
List of arff-files.
}
\description{
List datasets (arff-files) in a given path.
                                      #
List datasets (arff-files) in a given path.
}

