% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_experiments.R
\name{list_datasets_sorted}
\alias{list_datasets_sorted}
\title{List datasets (arff-files) in a given path sorted by their number of attributes.
                                      #
List datasets (arff-files) in a given path sorted by their number of attributes.}
\usage{
list_datasets_sorted(datapath)
}
\arguments{
\item{datapath}{A path.}
}
\value{
List of arff-files (sorted).
}
\description{
List datasets (arff-files) in a given path sorted by their number of attributes.
                                      #
List datasets (arff-files) in a given path sorted by their number of attributes.
}

