% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_experiments.R
\name{load_result}
\alias{load_result}
\title{Load results.
                                      #
Load results in a given folder (where they were saved using
\code{save_result}, given the name of the dataset and the name of
the used classifier).}
\usage{
load_result(datapath, dataset, classifier)
}
\arguments{
\item{datapath}{Directory with the data files.}

\item{dataset}{The name of the dataset.}

\item{classifier}{The name of the classifier.}
}
\value{
Nothing.
}
\description{
Load results.
                                      #
Load results in a given folder (where they were saved using
\code{save_result}, given the name of the dataset and the name of
the used classifier).
}

