% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{make_group_matrix}
\alias{make_group_matrix}
\title{Crate a results matrix from the analysis results of a dataset.}
\usage{
make_group_matrix(res, alpha = 0.05, full_tree = FALSE)
}
\arguments{
\item{res}{A results structure, i.e., the output from \code{analyze_dataset}.}

\item{alpha}{The significance level for filtering (showing) trees only above this threshold. The default is 0.05.}

\item{full_tree}{Return the full tree (also trees for which p < alpha). Default is \code{FALSE}.}
}
\value{
A matrix with the results.
}
\description{
Create a matrix summarising the results from the analysis of a
datast. The table contains groupings (denoted by letters) and the
degree, accuracy and p-value for each grouping.
}

