% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syn_data.R
\name{make_synthetic_dataset}
\alias{make_synthetic_dataset}
\title{Generate synthetic dataset.}
\usage{
make_synthetic_dataset(N = 500, mg2 = 0.3, seed = 42)
}
\arguments{
\item{N}{Number of samples in each class.}

\item{mg2}{Classwise means of second group (attibute 3)}

\item{seed}{Random seed. Default is 42.}
}
\description{
This function generates a synthetic dataset that corresponds to a
tree of the form
tree <- list(c(5, 1, 2), c(5, 3), c(0, 4))
}

