% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_syn_data.R
\name{mksigma}
\alias{mksigma}
\title{Generate covariance matrix.}
\usage{
mksigma(co, N)
}
\arguments{
\item{N}{Number of variables.}

\item{Correlation}{between the variables.}
}
\description{
Generate covariance matrix.
}
\keyword{internal}

