% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{permutation}
\alias{permutation}
\title{Permute the data matrix using the permutation tree.}
\usage{
permutation(tree, data)
}
\arguments{
\item{tree}{The permutation tree}

\item{data}{The data matrix}
}
\value{
A permutation of the indices of the rows in the data matrix
}
\description{
Permute the data matrix using the permutation tree.
}

