% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{permutedata}
\alias{permutedata}
\title{Use the permutation to obtain a new data matrix.}
\usage{
permutedata(data, perm)
}
\arguments{
\item{data}{The data matrix}

\item{perm}{The permutation}
}
\value{
The permuted data matrix
}
\description{
Use the permutation to obtain a new data matrix.
}

