% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{print_result_table_html}
\alias{print_result_table_html}
\title{Print the results table in HTML}
\usage{
print_result_table_html(res, alpha = 0.05, full_tree = FALSE)
}
\arguments{
\item{alpha}{The significance level for filtering (showing) trees only above this threshold. The default is 0.05.}

\item{full_tree}{Return the full tree (also trees for which p < alpha). Default is \code{FALSE}.}
}
\value{
Nothing.
}
\description{
Print the results table in HTML
}
\details{
#' @param res A results structure, i.e., the output from
    \code{analyze_dataset}.
}

