% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_experiments.R
\name{read_uci_dataset}
\alias{read_uci_dataset}
\title{Read UCI data set stored in arff format and transform all columns
to factors.}
\usage{
read_uci_dataset(datapath, dataset, factorize = FALSE, na.omit = TRUE,
  remove_constant = TRUE)
}
\arguments{
\item{datapath}{The path to the directory containing the datasets,
with trailing slash.}

\item{dataset}{The name of the dataset without file suffix.}

\item{factorize}{Should the columns be transformed to
factors. Default is FALSE}

\item{na.omit}{Omit NAs. Default is TRUE.}
}
\value{
The classification accuracy.
}
\description{
Read UCI data set stored in arff format and transform all columns
to factors.
}

