% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_experiments.R
\name{save_result}
\alias{save_result}
\title{Save results as rds file.
                                      #
Save results as rds file, automatically generating the filename based on the dataset name and the used classifier.}
\usage{
save_result(res, savepath)
}
\arguments{
\item{res}{The results structure.}

\item{savepath}{Path where to save the resutls.}
}
\value{
Nothing.
}
\description{
Save results as rds file.
                                      #
Save results as rds file, automatically generating the filename based on the dataset name and the used classifier.
}

