% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_find_singleton}
\alias{sid_find_singleton}
\title{Find attribute to remove from a tree.}
\usage{
sid_find_singleton(data_train, data_test, g_ut, g_s, id_cl, R = 10,
  classifier, parallel = TRUE)
}
\arguments{
\item{data_train}{Training data.}

\item{data_test}{Testing data.}

\item{g_ut}{List of untested singletons.}

\item{g_s}{List with already removed attributes as singletons.}

\item{id_cl}{The number of the class attribute.}

\item{R}{Number of replications. Default is 10.}

\item{classifier}{A classifier function, default is svm (from the e1071 package).}

\item{parallel}{Calculate in parallel (Boolean, default is \code{TRUE}).}
}
\value{
The attribute to remove.
}
\description{
Testing all possibilities, find the attribute to remove from the
grouping that impacts the goodness the least.
}

