% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_gen_surrogate}
\alias{sid_gen_surrogate}
\title{Generate a surrogate (permuted) dataset.}
\usage{
sid_gen_surrogate(tree, data_orig)
}
\arguments{
\item{tree}{The parametrisation of the permutation.}

\item{data_orig}{The data to permute.}
}
\value{
The permutation of data_orig using tree.
}
\description{
Generate a surrogate (permuted) dataset.
}

