% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_get_goodness}
\alias{sid_get_goodness}
\title{Get the goodness (accuracy) of a classifier.}
\usage{
sid_get_goodness(data_train, data_test, classifier = svm)
}
\arguments{
\item{data_train}{Training data.}

\item{data_test}{Testing data.}

\item{classifier}{A classifier function, default is svm (from the e1071 package).}
}
\value{
The accuracy of the classifier.
}
\description{
Get the goodness (accuracy) of a classifier.
}

