% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_group}
\alias{sid_group}
\title{Find a grouping from a sorted list of attributes.}
\usage{
sid_group(data_train, data_test, attr_rank, R = 10, classifier = svm,
  parallel = TRUE)
}
\arguments{
\item{data_train}{Training data.}

\item{data_test}{Testing data.}

\item{attr_rank}{Sorted list of attributes (from \code{sid_sort}).}

\item{R}{Number of replications. Default is 10.}

\item{classifier}{A classifier function, default is svm (from the e1071 package).}

\item{parallel}{Calculate in parallel (Boolean, default is \code{TRUE}).}
}
\value{
A list with the left and right parts of x split at the
    index si.
}
\description{
Find the best grouping from a sorted list of attributes and the goodness.
}

