% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_halve}
\alias{sid_halve}
\title{Calculate goodness for 2-group attribute sets.}
\usage{
sid_halve(data_train, data_test, attr_rank, R = 10, classifier = svm,
  parallel = parallel)
}
\arguments{
\item{data_train}{Training data.}

\item{data_test}{Testing data.}

\item{attr_rank}{Sorted list of attributes (from \code{sid_sort}).}

\item{R}{Number of replications. Default is 10.}

\item{classifier}{A classifier function, default is svm (from the e1071 package).}

\item{parallel}{Calculate in parallel (Boolean, default is \code{TRUE}).}
}
\value{
The attributes, sorted.
}
\description{
fter ranking, the attributes are ordered so that those belonging to
the same group are consecutive. Now calculate the accuracy for each
grouping formed by splitting this sorted list of attributes into a
tree with 2 groups, for each possible split (number of attributes -
1).
}

