% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_p}
\alias{sid_p}
\title{Calculate the p-values for a list of trees.}
\usage{
sid_p(data_train, data_test, tree_list, g0 = NULL, classifier = svm,
  Rmin = 250, Rmax = 500, alpha = 0.05, z = 2.6, parallel = TRUE)
}
\arguments{
\item{data_train}{Training data.}

\item{data_test}{Testing data.}

\item{tree_list}{List of trees.}

\item{g0}{The Baseline goodness, if known. Can be \code{NULL} (default).}

\item{classifier}{A classifier function, default is svm (from the e1071 package).}

\item{Rmin}{Ninimum number of replications. Default is 250.}

\item{Rmax}{Maximum number of replications. Default is 500.}

\item{alpha}{The confidence level at which we are evaluating statistical significance of the tree.}

\item{z}{Parameter for confidence band. Use 1.96 for 95 percent, 2.25 for 97.5 percent and 2.6 for 99 percent. Default is 2.6.}

\item{parallel}{Calculate in parallel (Boolean, default is \code{TRUE}).}
}
\value{
The treelist with the p-values added.
}
\description{
Calculate the p-value for a list of trees (groupings).
}

