% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_p_tree}
\alias{sid_p_tree}
\title{Calculate the p-value for a tree.}
\usage{
sid_p_tree(data_train, data_test, tree, g0 = NULL, classifier = svm,
  R = 100, histogram = FALSE, parallel = TRUE, return_raw = FALSE)
}
\arguments{
\item{data_train}{Training data.}

\item{data_test}{Testing data.}

\item{tree}{The tree}

\item{g0}{The Baseline goodness, if known. Can be \code{NULL} (default).}

\item{classifier}{A classifier function, default is svm (from the e1071 package).}

\item{R}{Number of replications. Default is 100.}

\item{histogram}{Plot histogram (Boolean, default is \code{FALSE}.}

\item{parallel}{Calculate in parallel (Boolean, default is \code{TRUE}).}

\item{return_raw}{Return raw data used in the p-value calculation (Boolean, default is \code{FALSE}.}
}
\value{
The p-value for the tree
}
\description{
Calculate the p-value for a tree (grouping).
}

