% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_part}
\alias{sid_part}
\title{Partition an array into two.}
\usage{
sid_part(x, si)
}
\arguments{
\item{x}{An array.}

\item{si}{Split index.}
}
\value{
A list with the left and right parts of x split at the
    index si.
}
\description{
Partition an array x into two lists at the index si.
}

