% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_prune_group}
\alias{sid_prune_group}
\title{Prune group.}
\usage{
sid_prune_group(tree, data, pg)
}
\arguments{
\item{tree}{The tree.}

\item{data}{The data with the attributes.}

\item{pg}{The group to prune.}
}
\value{
A valid tree.
}
\description{
Prune group (singleton or regular group).
}

