% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_prune_singletons}
\alias{sid_prune_singletons}
\title{Prune singletons from tree.}
\usage{
sid_prune_singletons(tree_res, data_train, data_test, classifier = svm,
  R = 100, Rmin = 250, Rmax = 500, z = 2.6, alpha = 0.05,
  parallel = TRUE)
}
\arguments{
\item{tree_res}{A tree results structure.}

\item{data_train}{Training data.}

\item{data_test}{Testing data.}

\item{classifier}{A classifier function, default is svm (from the e1071 package).}

\item{R}{Number of replications. Default is 100.}

\item{Rmin}{Ninimum number of replications for calculating p-values. Default is 250.}

\item{Rmax}{Maximum number of replications for calculating p-values. Default is 500.}

\item{z}{Parameter for confidence band. Use 1.96 for 95 percent, 2.25 for 97.5 percent and 2.6 for 99 percent. Default is 2.6.}

\item{alpha}{Significance level (default is 0.05).}

\item{parallel}{Calculate in parallel (Boolean, default is \code{TRUE}).}
}
\value{
The pruned tree results structure.
}
\description{
Prune singletons from tree.
}

