% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_prune_tree}
\alias{sid_prune_tree}
\title{Prune tree.}
\usage{
sid_prune_tree(tree_res, data_train, data_test, classifier = svm, R = 100,
  alpha = 0.05, parallel = TRUE)
}
\arguments{
\item{tree_res}{A tree results structure.}

\item{data_train}{Training data.}

\item{data_test}{Testing data.}

\item{classifier}{A classifier function, default is svm (from the e1071 package).}

\item{R}{Number of replications. Default is 100.}

\item{alpha}{Significance level (default is 0.05).}

\item{parallel}{Calculate in parallel (Boolean, default is \code{TRUE}).}
}
\value{
The pruned tree results structure.
}
\description{
Prune tree.
}

