% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_singletonize}
\alias{sid_singletonize}
\title{Singletonize the attributes in the vector x.}
\usage{
sid_singletonize(x, cl_id)
}
\arguments{
\item{x}{A vector with attribute numbers.}

\item{cl_id}{The number of the class attribute.}
}
\value{
A list where the attributes in x are singletons.
}
\description{
Singletonize the attributes in the vector x.
}

