% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_sort}
\alias{sid_sort}
\title{Rank attributes}
\usage{
sid_sort(data_train, data_test, R = 10, classifier = svm, parallel = TRUE)
}
\arguments{
\item{data_train}{Training data.}

\item{data_test}{Testing data.}

\item{R}{Number of replications. Default is 10.}

\item{classifier}{A classifier function, default is svm (from the e1071 package).}

\item{parallel}{Calculate in parallel (Boolean, default is \code{TRUE}).}
}
\value{
The attributes, sorted.
}
\description{
After ranking, the attributes are ordered so that those belonging
to the same group are consecutive.
}

