% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_tree_goodness}
\alias{sid_tree_goodness}
\title{Get the goodness (accuracy) of a tree.}
\usage{
sid_tree_goodness(data_train, data_test, tree, classifier, R = 10,
  parallel = TRUE, stats = FALSE)
}
\arguments{
\item{data_train}{Training data.}

\item{data_test}{Testing data.}

\item{classifier}{A classifier function, default is svm (from the e1071 package).}

\item{R}{Number of replications. Default is 10.}

\item{parallel}{Calculate in parallel (Boolean, default is \code{TRUE}).}

\item{stats}{Return statistics for the tree. Default is \code{FALSE}.}
}
\value{
The average goodness.
}
\description{
Get the goodness (accuracy) of a tree.
}

