% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_treeify}
\alias{sid_treeify}
\title{Create a valid tree from a list of attribute indices.}
\usage{
sid_treeify(tree, data)
}
\arguments{
\item{tree}{The tree.}

\item{data}{The data with the attributes.}
}
\value{
A valid tree.
}
\description{
Create a valid tree from a list of attribute indices by adding the
index of the class attribute to each group in the tree.
}

