% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sid_main_algorithm.R
\name{sid_untreeify}
\alias{sid_untreeify}
\title{Undo treeification.}
\usage{
sid_untreeify(tree, data)
}
\arguments{
\item{tree}{The tree.}

\item{data}{The data with the attributes.}
}
\value{
A valid tree.
}
\description{
Remove the class index from each group in the tree.
}

