% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_experiments.R
\name{split_dataset_ttv}
\alias{split_dataset_ttv}
\title{Split dataset into 3 parts: training, testing, validation (50/25/25).
                                      #
Split dataset into 3 parts: training, testing, validation (50/25/25).}
\usage{
split_dataset_ttv(dataset, classname = "class")
}
\arguments{
\item{dataset}{The dataset.}

\item{classname}{the name of the class in the dataset. Default is \code{class}.}
}
\value{
The dataset structure with two new elements (\code{data_train}, \code{data_test} and \code{data_validation}).
}
\description{
Split dataset into 3 parts: training, testing, validation (50/25/25).
                                      #
Split dataset into 3 parts: training, testing, validation (50/25/25).
}

