% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities_analysis.R
\name{tree_to_letters}
\alias{tree_to_letters}
\title{Represent a grouping (tree) using letters.}
\usage{
tree_to_letters(tree)
}
\arguments{
\item{tree}{A tree}
}
\value{
The tree represented using letters.
}
\description{
Represent a grouping (tree) using letters.
}

